+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/navigation/index.tsx
@@ -12,4 +12,4 @@
  * <http://www.gnu.org/licenses/lgpl.html>.
  *
  **/
-export { default } from './navigation-container'
+export { default } from './container'
